requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Async support
aiohttp==3.9.1
aiogram==3.3.0

# Environment variables
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0

# Date and time
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Optional: Maps integration
# geopy==2.4.1
# yandex-geocoder==2.0.0

